+++ Tests/Functional/Service/RegistrationManagerTest.php
@@ -5,8 +5,14 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\Tests\Functional\Service;
 
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
+use OliverKlee\Oelib\Authentication\FrontEndLoginManager;
 use OliverKlee\Oelib\Configuration\ConfigurationRegistry;
 use OliverKlee\Oelib\Configuration\DummyConfiguration;
+use OliverKlee\Oelib\Mapper\MapperRegistry;
+use OliverKlee\Oelib\Testing\TestingFramework;
+use OliverKlee\Seminars\FrontEnd\DefaultController;
+use OliverKlee\Seminars\Mapper\FrontEndUserMapper;
+use OliverKlee\Seminars\OldModel\LegacyEvent;
 use OliverKlee\Seminars\OldModel\LegacyRegistration;
 use OliverKlee\Seminars\Service\RegistrationManager;
 use OliverKlee\Seminars\Tests\Functional\Traits\LanguageHelper;
@@ -14,6 +20,7 @@ use OliverKlee\Seminars\Tests\Unit\Traits\EmailTrait;
 use OliverKlee\Seminars\Tests\Unit\Traits\MakeInstanceTrait;
 use TYPO3\CMS\Core\Mail\MailMessage;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
 /**
  * @covers \OliverKlee\Seminars\Service\RegistrationManager
@@ -25,19 +32,24 @@ final class RegistrationManagerTest extends FunctionalTestCase
     use MakeInstanceTrait;
 
     /**
-     * @var string
+     * @var non-empty-string
      */
     private const EMAIL_TEMPLATE_PATH = 'EXT:seminars/Resources/Private/Templates/Mail/e-mail.html';
 
     /**
-     * @var array<int, string>
+     * @var array<int, non-empty-string>
      */
     protected $testExtensionsToLoad = ['typo3conf/ext/oelib', 'typo3conf/ext/seminars'];
 
     /**
+     * @var TestingFramework
+     */
+    private $testingFramework;
+
+    /**
      * @var RegistrationManager
      */
-    private $subject = null;
+    private $subject;
 
     /**
      * @var DummyConfiguration
@@ -54,11 +66,15 @@ final class RegistrationManagerTest extends FunctionalTestCase
 
         $this->email = $this->createEmailMock();
 
+        $this->testingFramework = new TestingFramework('tx_seminars');
+
         $this->subject = RegistrationManager::getInstance();
     }
 
     protected function tearDown(): void
     {
+        $this->testingFramework->cleanUpWithoutDatabase();
+
         ConfigurationRegistry::purgeInstance();
         RegistrationManager::purgeInstance();
         // Purge the FIFO buffer of mocks
@@ -67,6 +83,47 @@ final class RegistrationManagerTest extends FunctionalTestCase
 
         parent::tearDown();
     }
+
+    private function getFrontEndController(): TypoScriptFrontendController
+    {
+        $controller = $GLOBALS['TSFE'];
+        if (!$controller instanceof TypoScriptFrontendController) {
+            throw new \RuntimeException('No FE present!', 1645868170);
+        }
+
+        return $controller;
+    }
+
+    /**
+     * @return positive-int user UID
+     */
+    private function createAndLogInUser(): int
+    {
+        $userUid = 1;
+        $this->importDataSet(__DIR__ . '/Fixtures/FrontEndUser.xml');
+        $this->logInUser($userUid);
+
+        return $userUid;
+    }
+
+    private function logInUser(int $uid): void
+    {
+        $user = MapperRegistry::get(FrontEndUserMapper::class)->find($uid);
+        FrontEndLoginManager::getInstance()->logInUser($user);
+    }
+
+    private function setUpFrontEndForRegistrationLink(): DefaultController
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/RegistrationAndLoginPage.xml');
+        $this->testingFramework->createFakeFrontEnd(1);
+        $controller = new DefaultController();
+        $controller->cObj = $this->getFrontEndController()->cObj;
+        $controller->conf = ['loginPID' => '2', 'registerPID' => '3'];
+
+        return $controller;
+    }
+
+    // Tests concerning notifyOrganizers
 
     /**
      * @test
@@ -85,5 +142,204 @@ final class RegistrationManagerTest extends FunctionalTestCase
 
         $expectedExpression = '/' . $this->getLanguageService()->getLL('label_vacancies') . ': 1\\n*$/';
         self::assertRegExp($expectedExpression, $this->getTextBodyOfEmail($this->email));
+    }
+
+    // Tests concerning getRegistrationLink
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithoutLoginAndPastEventReturnsEmptyString(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/PastEventWithVacancies.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertSame('', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithoutLoginWithFutureEventWithPassedDeadlineReturnsEmptyString(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/FutureEventWithPassedRegistrationDeadline.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertSame('', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithoutLoginAndPriceOnRequestReturnsEmptyString(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/EventWithPriceOnRequest.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertSame('', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithoutLoginAndFullyBookedEventReturnsEmptyString(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/FullyBookedEvent.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertSame('', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithoutLoginAndEventWithVacanciesReturnsLoginLink(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/EventWithVacancies.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertStringContainsString('/login', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithoutLoginAndEventWithUnlimitedVacanciesReturnsLoginLink(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/EventWithUnlimitedVacancies.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertStringContainsString('/login', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithoutLoginAndFullyBookedEventWithQueueEnabledReturnsLoginLink(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/FullyBookedEventWithQueue.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertStringContainsString('/login', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithLoginAndEventWithVacanciesReturnsLinkToRegistrationPage(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+        $this->createAndLogInUser();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/EventWithVacancies.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertStringContainsString('/registration', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithLoginAndEventWithVacanciesReturnsLinkWithEventUid(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+        $this->createAndLogInUser();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/EventWithVacancies.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertStringContainsString('%5Bseminar%5D=1', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithLoginAndFullyBookedEventReturnsEmptyString(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+        $this->createAndLogInUser();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/FullyBookedEvent.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertSame('', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithLoginAndEventWithUnlimitedVacanciesReturnsLinkWithEventUid(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+        $this->createAndLogInUser();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/EventWithUnlimitedVacancies.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertStringContainsString('%5Bseminar%5D=1', $result);
+    }
+
+    /**
+     * @test
+     */
+    public function getRegistrationLinkWithLoginAndFullyBookedEventWithQueueReturnsLinkWithEventUid(): void
+    {
+        $plugin = $this->setUpFrontEndForRegistrationLink();
+        $this->createAndLogInUser();
+
+        $this->importDataSet(__DIR__ . '/Fixtures/FullyBookedEventWithQueue.xml');
+        $event = LegacyEvent::fromUid(1);
+        self::assertInstanceOf(LegacyEvent::class, $event);
+
+        $result = $this->subject->getRegistrationLink($plugin, $event);
+
+        self::assertStringContainsString('%5Bseminar%5D=1', $result);
     }
 }